GH-12103 rename e2e helper classes
