+++ projects/storefrontapp-e2e-cypress/cypress/helpers/product-configurator.ts
