+++ projects/storefrontapp-e2e-cypress/cypress/integration/accessibility/product-configuration-tabbing.flaky-e2e-spec.ts
@@ -1,8 +1,8 @@
 import { verifyTabbingOrder } from '../../helpers/accessibility/tabbing-order';
 import { tabbingOrderConfig as tabConfig } from '../../helpers/accessibility/tabbing-order.config';
-import * as configurationOverview from '../../helpers/product-configuration-overview';
+import * as configurationOverview from '../../helpers/product-configurator-overview';
 import * as configurationVc from '../../helpers/product-configurator-vc';
-import * as configuration from '../../helpers/product-configuration';
+import * as configuration from '../../helpers/product-configurator';
 /**
  * This suite is marked as flaky due to performance (synchronization) issues on
  * https://spartacus-devci767.eastus.cloudapp.azure.com:9002 that we analyze in
