+++ projects/storefrontapp-e2e-cypress/cypress/integration/mobile/product-configuration-mobile.flaky-e2e-spec.ts
@@ -1,4 +1,4 @@
-import * as configuration from '../../helpers/product-configuration';
+import * as configuration from '../../helpers/product-configurator';
 import * as configurationVc from '../../helpers/product-configurator-vc';
 import { formats } from '../../sample-data/viewports';
