+++ projects/storefrontapp-e2e-cypress/cypress/integration/regression/product-configuration-cart.flaky-e2e-spec.ts
@@ -1,9 +1,9 @@
 import * as cart from '../../helpers/cart';
 import * as login from '../../helpers/login';
-import * as configuration from '../../helpers/product-configuration';
-import * as configurationOverview from '../../helpers/product-configuration-overview';
+import * as configuration from '../../helpers/product-configurator';
+import * as configurationOverview from '../../helpers/product-configurator-overview';
 import * as configurationVc from '../../helpers/product-configurator-vc';
-import * as configurationOverviewVc from '../../helpers/product-configurator-vc-overview';
+import * as configurationOverviewVc from '../../helpers/product-configurator-overview-vc';
 import * as configurationCart from '../../helpers/product-configurator-cart';
 import * as configurationCartVc from '../../helpers/product-configurator-cart-vc';
 import * as productSearch from '../../helpers/product-search';
